CCCP-488, refactor: rewrite in dyn Trait
