+++ Cargo.toml
@@ -71,7 +71,6 @@ alloy = { version = "0.8", features = [
 k256 = "0.13.1"
 sha3 = "0.10.8"
 aes-gcm = "0.10.3"
-pbkdf2 = "0.12.2"
 
 aws-sdk-kms = "1.54.0"
 aws-config = { version = "1.5.13", features = ["behavior-version-latest"] }
