+++ client/Cargo.toml
@@ -37,7 +37,6 @@ eyre = { workspace = true }
 aws-sdk-kms = { workspace = true }
 aes-gcm = { workspace = true }
 rand = { workspace = true }
-pbkdf2 = { workspace = true, features = ["simple"] }
 
 # subs
 sc-keystore = { workspace = true }
