+++ periodic/src/psbt_signer.rs
@@ -8,7 +8,7 @@ use alloy::{
 use br_client::{
 	btc::{
 		handlers::XtRequester,
-		storage::keypair::{KeypairManager, KeypairStorage},
+		storage::keypair::{KeypairStorage, KeypairStorageT},
 	},
 	eth::EthClient,
 };
@@ -30,19 +30,18 @@ use tokio_stream::StreamExt;
 const SUB_LOG_TARGET: &str = "psbt-signer";
 
 /// The essential task that submits signed PSBT's.
-pub struct PsbtSigner<F, P, T, K>
+pub struct PsbtSigner<F, P, T>
 where
 	F: TxFiller<AnyNetwork> + WalletProvider<AnyNetwork>,
 	P: Provider<T, AnyNetwork>,
 	T: Transport + Clone,
-	K: KeypairManager + Send + Sync + 'static,
 {
 	/// The Bifrost client.
 	pub client: Arc<EthClient<F, P, T>>,
 	/// The unsigned transaction message sender.
 	xt_request_sender: Arc<XtRequestSender>,
 	/// The public and private keypair local storage.
-	keypair_storage: Arc<RwLock<KeypairStorage<K>>>,
+	keypair_storage: KeypairStorage,
 	/// The migration sequence.
 	migration_sequence: Arc<RwLock<MigrationSequence>>,
 	/// The Bitcoin network.
@@ -51,18 +50,17 @@ where
 	schedule: Schedule,
 }
 
-impl<F, P, T, K> PsbtSigner<F, P, T, K>
+impl<F, P, T> PsbtSigner<F, P, T>
 where
 	F: TxFiller<AnyNetwork> + WalletProvider<AnyNetwork>,
 	P: Provider<T, AnyNetwork>,
 	T: Transport + Clone,
-	K: KeypairManager + Send + Sync + 'static,
 {
 	/// Instantiates a new `PsbtSigner` instance.
 	pub fn new(
 		client: Arc<EthClient<F, P, T>>,
 		xt_request_sender: Arc<XtRequestSender>,
-		keypair_storage: Arc<RwLock<KeypairStorage<K>>>,
+		keypair_storage: KeypairStorage,
 		migration_sequence: Arc<RwLock<MigrationSequence>>,
 		btc_network: Network,
 	) -> Self {
@@ -124,7 +122,7 @@ where
 		};
 
 		let mut psbt = unsigned_psbt.clone();
-		if self.keypair_storage.read().await.0.sign_psbt(&mut psbt) {
+		if self.keypair_storage.sign_psbt(&mut psbt).await {
 			let signed_psbt = psbt.serialize();
 
 			if self
@@ -205,12 +203,11 @@ where
 }
 
 #[async_trait::async_trait]
-impl<F, P, T, K> XtRequester<F, P, T> for PsbtSigner<F, P, T, K>
+impl<F, P, T> XtRequester<F, P, T> for PsbtSigner<F, P, T>
 where
 	F: TxFiller<AnyNetwork> + WalletProvider<AnyNetwork>,
 	P: Provider<T, AnyNetwork>,
 	T: Transport + Clone,
-	K: KeypairManager + Send + Sync + 'static,
 {
 	fn xt_request_sender(&self) -> Arc<XtRequestSender> {
 		self.xt_request_sender.clone()
@@ -222,12 +219,11 @@ where
 }
 
 #[async_trait::async_trait]
-impl<F, P, T, K> PeriodicWorker for PsbtSigner<F, P, T, K>
+impl<F, P, T> PeriodicWorker for PsbtSigner<F, P, T>
 where
 	F: TxFiller<AnyNetwork> + WalletProvider<AnyNetwork>,
 	P: Provider<T, AnyNetwork>,
 	T: Transport + Clone,
-	K: KeypairManager + Send + Sync + 'static,
 {
 	fn schedule(&self) -> Schedule {
 		self.schedule.clone()
