+++ relayer/src/commands/migrate_keystore_cmd.rs
@@ -1,5 +1,6 @@
-use aws_sdk_kms::primitives::Blob;
-use br_client::btc::storage::keypair::{KeypairManager, KeypairStorage, KeypairStorageKind};
+use br_client::btc::storage::keypair::{
+	KeypairStorage, KeypairStorageT, KmsKeypairStorage, PasswordKeypairStorage,
+};
 use br_primitives::{
 	cli::Configuration,
 	constants::{cli::DEFAULT_KEYSTORE_PATH, errors::INVALID_BITCOIN_NETWORK},
@@ -8,14 +9,11 @@ use br_primitives::{
 use clap::Parser;
 use miniscript::bitcoin::Network;
 use sc_cli::Error as CliError;
-use sc_keystore::Keystore;
 use secrecy::SecretString;
-use secrecy::Zeroize;
 use sp_application_crypto::{
 	ecdsa::{AppPair, AppPublic},
 	ByteArray,
 };
-use sp_core::testing::ECDSA;
 use std::{process::Command, sync::Arc};
 
 use crate::cli::{MAINNET_CONFIG_FILE_PATH, TESTNET_CONFIG_FILE_PATH};
@@ -76,7 +74,7 @@ impl MigrateKeystoreCmd {
 				let aws_client = Arc::new(aws_sdk_kms::Client::new(
 					&aws_config::load_defaults(aws_config::BehaviorVersion::latest()).await,
 				));
-				KeypairStorage::new(KeypairStorageKind::new_kms(
+				KeypairStorage::new(KmsKeypairStorage::new(
 					keystore_path.clone(),
 					network,
 					key_id.clone(),
@@ -84,14 +82,14 @@ impl MigrateKeystoreCmd {
 				))
 			} else {
 				password = keystore_config.password.clone();
-				KeypairStorage::new(KeypairStorageKind::new_password(
+				KeypairStorage::new(PasswordKeypairStorage::new(
 					keystore_path.clone(),
 					network,
 					password.clone(),
 				))
 			}
 		} else {
-			KeypairStorage::new(KeypairStorageKind::new_password(
+			KeypairStorage::new(PasswordKeypairStorage::new(
 				DEFAULT_KEYSTORE_PATH.to_string(),
 				network,
 				None,
@@ -100,23 +98,11 @@ impl MigrateKeystoreCmd {
 
 		// 2. Load the keys from the current (old) keystore.
 		if self.version == 1 {
-			match old_keystore.0 {
-				KeypairStorageKind::Password(ref mut storage) => {
-					storage.inner.load_v1(self.round, password.clone());
-				},
-				KeypairStorageKind::Kms(_) => panic!("KMS keystore is not supported for version 1"),
-			}
+			old_keystore.load_v1(self.round, password.clone()).await;
 		} else {
-			old_keystore.0.load(self.round);
+			old_keystore.load(self.round).await;
 		}
-		let keys = match old_keystore.0.clone() {
-			KeypairStorageKind::Password(storage) => {
-				storage.inner.db().keys(ECDSA).expect("Failed to load keys")
-			},
-			KeypairStorageKind::Kms(storage) => {
-				storage.inner.db().keys(ECDSA).expect("Failed to load keys")
-			},
-		};
+		let keys = old_keystore.keys().await;
 
 		// 3. Backup the keystore.
 		let backup_path = format!("{}_backup", keystore_path);
@@ -133,133 +119,42 @@ impl MigrateKeystoreCmd {
 			let aws_client = Arc::new(aws_sdk_kms::Client::new(
 				&aws_config::load_defaults(aws_config::BehaviorVersion::latest()).await,
 			));
-			KeypairStorage::new(KeypairStorageKind::new_kms(
+			KeypairStorage::new(KmsKeypairStorage::new(
 				keystore_path.clone(),
 				network,
 				key_id.clone(),
 				aws_client,
 			))
 		} else {
-			KeypairStorage::new(KeypairStorageKind::new_password(
+			KeypairStorage::new(PasswordKeypairStorage::new(
 				keystore_path.clone(),
 				network,
 				self.new_password.clone(),
 			))
 		};
-		new_keystore.0.load(self.round);
+		new_keystore.load(self.round).await;
 
 		// 5. Insert the keys to the new keystore.
 		for key in keys {
-			match old_keystore.0 {
-				KeypairStorageKind::Password(ref storage) => {
-					if self.version == 1 {
-						match storage.inner.db().key_pair::<AppPair>(
-							&AppPublic::from_slice(&key).expect("Failed to get key pair"),
-						) {
-							Ok(pair) => {
-								if let Some(pair) = pair {
-									self.insert_key(
-										&new_keystore.0,
-										&key,
-										pair.into_inner().seed().as_slice(),
-									)
-									.await;
-								} else {
-									panic!("Failed to get key pair");
-								}
-							},
-							Err(err) => {
-								panic!("Failed to get key pair: {:?}", err)
-							},
-						}
-					} else {
-						match storage.inner.db().raw_keystore_value::<AppPair>(
-							&AppPublic::from_slice(&key).expect("Failed to get key pair"),
-						) {
-							Ok(value) => {
-								if let Some(value) = value {
-									let decoded =
-										hex::decode(value.as_bytes()).expect("Failed to decode");
-									let mut seed = if storage.secret.is_some() {
-										storage.decrypt_key(&decoded)
-									} else {
-										decoded
-									};
-									self.insert_key(&new_keystore.0, &key, &seed).await;
-									seed.zeroize();
-								} else {
-									panic!("Failed to get key pair");
-								}
-							},
-							Err(err) => {
-								panic!("Failed to get key pair: {:?}", err)
-							},
-						}
-					}
-				},
-				KeypairStorageKind::Kms(ref storage) => {
-					match storage.inner.db().raw_keystore_value::<AppPair>(
+			if self.version == 1 {
+				let pair = old_keystore
+					.db()
+					.await
+					.key_pair::<AppPair>(
 						&AppPublic::from_slice(&key).expect("Failed to get key pair"),
-					) {
-						Ok(value) => {
-							if let Some(value) = value {
-								let mut seed = storage
-									.decrypt_key(
-										&hex::decode(value.as_bytes()).expect("Failed to decode"),
-									)
-									.await;
-								self.insert_key(&new_keystore.0, &key, &seed).await;
-								seed.zeroize();
-							} else {
-								panic!("Failed to get key pair");
-							}
-						},
-						Err(err) => {
-							panic!("{:?}", err)
-						},
-					}
-				},
+					)
+					.expect("Failed to get key pair")
+					.expect("Failed to get key pair");
+				new_keystore.insert_key(&key, pair.into_inner().seed().as_slice()).await;
+			} else {
+				let value =
+					old_keystore.raw_keystore_value(&key).await.expect("Failed to get key pair");
+				let decoded_value = hex::decode(value.as_bytes()).expect("Failed to decode");
+				let decrypted_value = old_keystore.decrypt_key(&decoded_value).await;
+				new_keystore.insert_key(&key, &decrypted_value).await;
 			}
 		}
 
 		Ok(())
-	}
-
-	async fn insert_key(&self, keystore: &KeypairStorageKind, key: &[u8], value: &[u8]) {
-		match keystore {
-			KeypairStorageKind::Password(ref storage) => {
-				let value = if storage.secret.is_some() {
-					storage.encrypt_key(value)
-				} else {
-					value.to_vec()
-				};
-
-				storage
-					.inner
-					.db()
-					.insert(ECDSA, &hex::encode(value), key)
-					.expect("Failed to insert key");
-			},
-			KeypairStorageKind::Kms(ref storage) => {
-				let encrypted_key = storage
-					.client
-					.encrypt()
-					.key_id(storage.key_id.clone())
-					.plaintext(Blob::new(value))
-					.send()
-					.await
-					.expect("Failed to encrypt");
-
-				storage
-					.inner
-					.db()
-					.insert(
-						ECDSA,
-						&hex::encode(encrypted_key.ciphertext_blob.unwrap().as_ref()),
-						key,
-					)
-					.expect("Failed to insert key");
-			},
-		}
 	}
 }
